Replace convergence ratchet with stateless source policy checks
