+++ .github/workflows/cut-release.yml
@@ -58,7 +58,6 @@ jobs:
         uses: actions/checkout@v7.0.1
         with:
           ref: main
-          fetch-depth: 0
           # Check out with the App token so the later commit and tag pushes are
           # authenticated as the App and therefore trigger release.yml.
           token: ${{ steps.app-token.outputs.token }}
@@ -118,8 +117,8 @@ jobs:
         run: |
           set -euo pipefail
           python3 scripts/check-doc-anchors.py
-          python3 -m unittest -v scripts/test_convergence_ratchet.py
-          python3 scripts/convergence-ratchet.py
+          python3 -m unittest -v scripts/test_check_source_policy.py
+          python3 scripts/check-source-policy.py
           python3 -m unittest -v scripts/test_check_golden_coverage_floors.py
           python3 scripts/check-golden-coverage-floors.py
           python3 scripts/check-public-api-ledger.py --self-test
