+++ AGENTS.md
@@ -20,7 +20,7 @@ Test placement policy:
 - Semantic test submodules such as `foo/tests/parsing.rs` are allowed. Never use numbered names such as `tests_1.rs`.
 - Integration tests live under `src/integration_tests.rs` or a semantic `integration_tests/` tree. Use them for codec-wide decode, encode, or round-trip; multi-module composition; the crate facade; or cross-module admission, validation, or fidelity work with no single owner. "Unclear" is not integration by default.
 - Test support lives under `src/test_support.rs` or a semantic `test_support/` tree. Single-owner helpers stay with that owner. Shared golden and round-trip helpers use `cadmpeg-test-support`. Do not add a new support crate. Test support is not exempt from size convergence.
-- Golden tests live under `src/golden_tests.rs` or a semantic `golden_tests/` tree. They are excluded from test-line debt. Do not regenerate snapshots during this migration.
+- Golden tests live under `src/golden_tests.rs` or a semantic `golden_tests/` tree. They are excluded from test-file size limits. Do not regenerate snapshots during test moves.
 - A crate root may expose only three test-only entry points: `golden_tests`, `integration_tests`, and `test_support`. Unit-test modules are declared by their production owners.
 - Do not split production only because a cohesive test suite is large. Split production only when an independent review finds a real responsibility seam.
 - Test moves preserve behavior. Do not change assertions, expected values, byte literals, tolerances, snapshots, test names, or ignored status.
@@ -53,14 +53,14 @@ Build and test operations:
 - sccache: evaluated, no disk or speed win in this workspace — do not re-add.
 - Bare tolerance literals: when touching a line with bare `1e-6` / `1e-7` / `1e-8` / `1e-9` / `1e-10` / `1e-11` / `1e-12`, hoist to a module-local named constant with intent in the name (`EPS_SAME_POINT`, `EPS_ORTHO`, etc.). Never collapse distinct tolerance values. Orthonormality thresholds are acceptance gates; values differ per codec.
 - Layout-table constants generation is in: `UPDATE_LAYOUT_CODE=1 cargo test -p cadmpeg --test layout_tables` writes `src/layout.rs` in each owning crate. Generation of parsing functions stays out. Tables remain the oracle; `cargo test -p cadmpeg --test layout_tables` checks arithmetic, citations, and that checked-in constants match the emitter.
-- Convergence ratchet: `scripts/convergence-ratchet.py` against `docs/convergence-ledger.toml`. Counts may only fall; update the ledger in the same commit as a decrease. A deliberate increase raises the ceiling and records a reason under `[reasons]`; `check` requires counts to equal ceilings; `--update` records decreases. CI supplies `--base` with the PR base or push-before revision and rejects increases without a new reason. Local checks default to `HEAD`. Do not store a measurement commit in the ledger; squash merges replace branch history. All metrics converge to `[targets]`. `from_endian_bytes` counts unapproved `from_le_bytes` / `from_be_bytes` calls in production. A `// endian-exception: reconstructed-scalar` or `// endian-exception: packed-color-order` comment admits exactly one call on the next line. Use these only for reconstructed byte representations or in-memory color ordering, never a standard-width file read. Stale or unknown exception comments fail the check. Convergence keys include `crate_root_tests_rs`, `path_test_includes`, `test_line_debt`, `production_line_debt`, `codec_error_malformed_format`, `loss_note_struct_literals`, `bare_tolerance_literals`, and `nonliteral_vec_repeat`. `[ceilings]` stores current measured debt. Text and JSON output list contributing paths. `bare_tolerance_literals` counts `1e-6` / `1e-7` / `1e-8` / `1e-9` / `1e-10` / `1e-11` / `1e-12` occurrences only; it does not see naming or collapse of distinct values.
+- Source policy: `python3 scripts/check-source-policy.py` checks the current source without Git history or a ledger. Findings contain rule, file, line, and explanation; any violation fails. Rules and lexical limits are in `docs/source-policy.md`. Endian exceptions require a standalone `// endian-exception: reconstructed-scalar` or `// endian-exception: packed-color-order` comment immediately before exactly one call. These admit reconstructed scalar bytes or in-memory color ordering, never ordinary standard-width file reads. Stale and unknown exceptions fail. Policy changes edit the rule and its tests, not a budget.
 - Golden fixture floors live in `docs/golden-coverage-floors.toml` and may only rise when fixtures are added; they must never silently fall.
 - Shared golden harness: `cadmpeg_test_support::golden`. Frozen-file codecs use `Harness::check`. Code-built inputs (nx, inventor, sat) use `Harness::check_inputs`.
 - NX goldens are code-built and pin flate2 zlib bytes. Regenerate with `UPDATE_GOLDEN=1 cargo test-fast golden` (workspace). A `-p cadmpeg-codec-nx` build can pick a different flate2 backend and rewrite snapshots.
 - NX in-tree arena coverage floors stay in `cadmpeg-codec-nx` golden tests; the TOML floors are file counts only.
 - A public API change regenerates the affected crate's snapshot under `docs/api-baseline/` in the same commit; `git diff docs/api-baseline/` is the change record. `docs/public-api-ledger.toml` holds only the header (`baseline_commit`, `api_baseline_dir`, `measured_at`); do not add `[[change]]` rows.
 - `scripts/perf-baseline.sh` is the decode/build timing harness for hot-path changes.
-- Checked allocation for parsed counts: prefer `DecodeContext::alloc_filled` (charges collection items) or `cadmpeg_core::decode::alloc_filled` (reserve-only) over `vec![value; parsed_count]`. The ratchet key `nonliteral_vec_repeat` tracks remaining sites.
+- Checked allocation for parsed counts: prefer `DecodeContext::alloc_filled` (charges collection items) or `cadmpeg_core::decode::alloc_filled` (reserve-only) over `vec![value; parsed_count]`. The source-policy rule `unchecked_vec_repeat` rejects unapproved repeats.
 - Single-codec CLI rebuild: `cargo build -p cadmpeg --no-default-features --features <id>` where `<id>` is a registry feature (`fcstd`, `f3d`, `inventor`, `sldprt`, `catia`, `creo`, `nx`, `rhino`, `step`, `iges`, `sat`). Release and default binaries keep all eleven codecs.
 - Bounded reads go through `cadmpeg_core::decode::View` (typed LE/BE reads, `read_counted`, `req_*`, offset helpers `u16_le_at` / `u32_be_at` / …). `cadmpeg_core::cursor::Cursor` is deleted. `cadmpeg_core::{le,be,read}` are deleted (superseded by View). Keep `cadmpeg_core::decode::bounded_len`. Do not flatten `View` to `&[u8]` at archive entry maps when SpaceId must survive. Prefer `?` on `req_*` via a codec-local `From<ParseError>`; do not `map_err` locations away. Sequential scanners keep a live `View`. Offset scanners over a retained payload use `View::u16_le_at(bytes, off)` (and the matching width/endian helper), not `T::from_le_bytes` / `from_be_bytes` on a window slice. Do not add new `from_le_bytes` / `from_be_bytes` on contiguous file windows. Exceptions: magic fourccs (`u32::from_le_bytes(*b"…")`), deliberately reordered or non-contiguous bytes, and packing of already-copied in-memory arrays that are not a decode window.
 - Codec crate roots are facades: export the codec, supported encoder/options, and
