--- docs/convergence-ledger.toml
@@ -1,38 +1,0 @@
-# Convergence ratchet ceilings. Counts may only fall.
-# A decrease updates this file in the same commit. A deliberate increase
-# requires a reason under [reasons] for every raised key.
-# All metrics converge to [targets]. Git records measurement provenance.
-filter = "legacy metrics: production .rs under crates/**/src via is_production_rs; exclude test, test-support, golden, integration, and bench path segments; exclude files named tests.rs or *test*.rs; lexically mask Rust comments and literals, then strip cfg(test)-attributed items with blank-preserving elision. bare tolerance literals count at use sites; named const/static numeric initializers are declarations and are excluded. nonliteral vec repeats exclude literal sizes and sizes derived from an existing collection's admitted len(). from_endian_bytes counts unapproved calls in that same crates/**/src glob. A preceding endian-exception comment admits one call for reconstructed-scalar or packed-color-order. Placement metrics: scan crates/**/*.rs by ownership, structural entry points, standard mod resolution, and test-only #[path] ancestry; golden_tests files stay out of test-line debt. production_line_debt reuses is_production_rs and elides cfg(test) items without blank placeholders. All metrics converge to their [targets]."
-
-[kinds]
-from_endian_bytes = "convergence"
-codec_error_malformed_format = "convergence"
-loss_note_struct_literals = "convergence"
-bare_tolerance_literals = "convergence"
-nonliteral_vec_repeat = "convergence"
-crate_root_tests_rs = "convergence"
-path_test_includes = "convergence"
-test_line_debt = "convergence"
-production_line_debt = "convergence"
-
-[targets]
-from_endian_bytes = 0
-codec_error_malformed_format = 0
-loss_note_struct_literals = 0
-bare_tolerance_literals = 0
-nonliteral_vec_repeat = 0
-crate_root_tests_rs = 0
-path_test_includes = 0
-test_line_debt = 0
-production_line_debt = 0
-
-[ceilings]
-from_endian_bytes = 0
-codec_error_malformed_format = 0
-loss_note_struct_literals = 0
-bare_tolerance_literals = 0
-nonliteral_vec_repeat = 0
-crate_root_tests_rs = 0
-path_test_includes = 0
-test_line_debt = 0
-production_line_debt = 0
